lots of changes

+++ TheRobot/GPSpython.py
@@ -20,6 +20,82 @@ def distance(lat1, lon1, lat2, lon2):
 
   return (Base * 1000)
  
+def getInitialBearing():
+    initialBearing = 0.0
+    
+    #get the initial cart position
+    cartLatInitial = '43.816142300' #gpsd.fix.latitude
+    cartLongInitial = '-111.784585100' #gpsd.fix.longitude
+    
+    os.system('python /home/pi/motor.py f 1 &')
+    time.sleep(2.5)  #drive forward for 1 meter
+    os.system('python /home/pi/motor.py f 0 &')
+    
+    #get the carts position
+    cartLat = '43.816142399' #gpsd.fix.latitude
+    cartLong = '-111.784585100' #gpsd.fix.longitude
+
+    # convert to radians:
+    g2r = pi/180
+    lat1r = float(cartLatInitial) * g2r
+    lat2r = float(cartLat) * g2r
+    lon1r = float(cartLongInitial) * g2r
+    lon2r = float(cartLong) * g2r
+    dlonr = lon2r - lon1r
+    y = sin(dlonr) * cos(lat2r)
+    x = cos(lat1r) * sin(lat2r) - sin(lat1r) * cos(lat2r) * cos(dlonr)
+
+    # compute bearning and convert back to degrees:
+    initialBearing = atan2(y, x) / g2r
+
+    #if initialBearing < 0:
+      #initialBearing = initialBearing + 360
+
+    print 'Bearing: ' + str(initialBearing)
+    
+    return initialBearing 
+ 
+def getCurrentBearing(latCurrent, longCurrent, latDest, longDest):
+    currentBearing = 0.0
+    
+    # convert to radians:
+    g2r = pi/180
+    lat1r = float(latCurrent) * g2r
+    lat2r = float(latDest) * g2r
+    lon1r = float(longCurrent) * g2r
+    lon2r = float(longDest) * g2r
+    dlonr = lon2r - lon1r
+    y = sin(dlonr) * cos(lat2r)
+    x = cos(lat1r) * sin(lat2r) - sin(lat1r) * cos(lat2r) * cos(dlonr)
+
+    # compute bearning and convert back to degrees:
+    currentBearing = atan2(y, x) / g2r
+
+    #if currentBearing < 0:
+      #currentBearing = currentBearing + 360
+
+    print 'Bearing: ' + str(currentBearing)
+    
+    return currentBearing
+ 
+def turnTowardDestination(bearingToTurn):
+    degreesToTurn = abs(bearingToTurn)
+    if degreesToTurn > 180:
+        degreesToTurn = degreesToTurn - 180
+    sleeptime = degreesToTurn/280
+    if ((bearingToTurn > 0 and bearingToTurn <= 180) or (bearingToTurn > -360 and bearingToTurn <= -180)):
+        #turn right
+        print "turn right"
+        os.system('python /home/pi/motor.py r 1 &')
+        time.sleep(sleeptime)
+        os.system('python /home/pi/motor.py r 0 &')
+    elif ((bearingToTurn < 0 and bearingToTurn > -180) or (bearingToTurn > 180 and bearingToTurn < 360)):
+        #turn left
+        print "turn left"
+        os.system('python /home/pi/motor.py l 1 &')
+        time.sleep(sleeptime)
+        os.system('python /home/pi/motor.py l 0 &')
+ 
 class GpsPoller(threading.Thread):
   def __init__(self):
     threading.Thread.__init__(self)
@@ -43,16 +119,10 @@ if __name__ == '__main__':
     #print gpsd.fix.latitude,', ',gpsd.fix.longitude,'  Time: ',gpsd.utc
 
     os.system('clear')
-    #d = distance(gpsd.fix.latitude, gpsd.fix.longitude, sys.argv[1], sys.argv[2])
-    #d = distance(43.816142300, -111.784585100, 43.816136700, -111.784569000)
-    #d = distance(43.816142300, -111.784585100, float(sys.argv[1]), float(sys.argv[2]))
-    #f = 'f'
-    #os.system('python /home/pi/motortwo.py '+ f + ' ' + str(d) + ' &')
-    #if d > 1:
-    #  print 'distance ' , d
-    #else:
-    #  print 'in range'
-  
+    
+    #get the initial bearing by moving forward one meter
+    currentBearing = getInitialBearing()
+    
     following = 1;
     numRecords = 0
     while (int(numRecords) == 0):
@@ -69,16 +139,23 @@ if __name__ == '__main__':
       destinationLong = os.popen('mysql -B --disable-column-names --user=shane --password=password webservice -e \"SELECT longitude FROM gps WHERE gpsID=\'' + str(gpsMinId) + '\'";').read()
       destinationLat = os.popen('mysql -B --disable-column-names --user=shane --password=password webservice -e \"SELECT latitude FROM gps WHERE gpsID=\'' + str(gpsMinId) + '\'";').read()
     
+      #get the carts position
+      cartLat = '43.816142300' #gpsd.fix.latitude
+      cartLong = '-111.784585100' #gpsd.fix.longitude
+    
       #calculate direction to destination
+      bearingToDestination = getBearing(cartLat, cartLong, destinationLat, destinationLong)
+
+      #determine how far to turn and in what direction
+      bearingToTurn = bearingToDestination - currentBearing
+      
       #turn towards destination
+      turnTowardDestination(bearingToTurn)
+      currentBearing = bearingToDestination
     
       #initialize variables
       moving = False
       destinationReached = False
-    
-      #get the carts position
-      cartLat = '43.816142300' #gpsd.fix.latitude
-      cartLong = '-111.784585100' #gpsd.fix.longitude
     
       #get the players position
       gpsMaxId = os.popen('mysql -B --disable-column-names --user=shane --password=password webservice -e "SELECT max(gpsID) FROM gps ";').read()
@@ -87,6 +164,8 @@ if __name__ == '__main__':
       
       distanceToPlayer = distance(float(cartLat), float(cartLong), float(playerLat), float(playerLong))
       print 'distanceToPlayer = ' , distanceToPlayer
+      previousCartLat = cartLat
+      previousCartLong = cartLong
     
       #currently facing the right direction to go forwards toward destination 
       while (distanceToPlayer > 3 and (not destinationReached) and (int(following) == 1)):
@@ -103,6 +182,28 @@ if __name__ == '__main__':
         #get the carts position
         cartLat = '43.816142300' #gpsd.fix.latitude
         cartLong = '-111.784585100' #gpsd.fix.longitude
+        
+        #update current bearing to ensure that we are still on course to destination
+        if ((previousCartLat != cartLat) and (previousCartLong != cartLong)):
+            currentBearing = getBearing(previousCartLat, previousCartLong, cartLat, cartLong)
+        bearingToDestination = getBearing(cartLat, cartLong, destinationLat, destinationLong)
+        bearingToTurn = bearingToDestination - currentBearing
+        if (bearingToTurn != 0):
+            #correct the direction to be heading straight for goal
+            print "correcting direction"
+            
+            #stop moving forward so that we can perform the turn
+            os.system('python /home/pi/motor.py f 0 &')
+            #turn towards destination
+            turnTowardDestination(bearingToDestination)
+            currentBearing = bearingToDestination
+            
+            #update cart position
+            cartLat = '43.816142300' #gpsd.fix.latitude
+            cartLong = '-111.784585100' #gpsd.fix.longitude
+            
+            #resume moving forward
+            os.system('python /home/pi/motor.py f 1 &')
     
         #calculate distance to player
         distanceToPlayer = distance(float(cartLat), float(cartLong), float(playerLat), float(playerLong))
@@ -143,9 +244,10 @@ if __name__ == '__main__':
           #update following
           following = os.popen('mysql -B --disable-column-names --user=shane --password=password webservice -e "SELECT pinStatus FROM pinstatus WHERE pinNumber=''5''";').read() 
           print 'following = ' , following 
-        #delete all records in database except last one
-        print 'deleting all records in database except last one'
-        os.popen('mysql -B --disable-column-names --user=shane --password=password webservice -e "DELETE FROM gps where gpsID NOT IN ( SELECT * FROM (SELECT MAX(gpsID) FROM gps) AS X)";').read()      
+        if (int(following) == 1):
+          #delete all records in database except last one
+          print 'deleting all records in database except last one'
+          os.popen('mysql -B --disable-column-names --user=shane --password=password webservice -e "DELETE FROM gps where gpsID NOT IN ( SELECT * FROM (SELECT MAX(gpsID) FROM gps) AS X)";').read()      
     
       if (destinationReached and int(following) == 1):
         print 'destination reached'
